works now

+++ ffmpeg_wrappers/folder_convert.py
@@ -5,7 +5,7 @@
 #output: converted files in "converted" folder or in specified output folder (if provided)
 
 import sys
-from os import listdir,system # ().: listdir;  ().path.: dirname,splitext,join,isfile,abspath
+from os import listdir,system,mkdir # ().: listdir;  ().path.: dirname,splitext,join,isfile,abspath
 from os.path import dirname,splitext,join,isfile,abspath
 import glob
 from shutil import copy2
@@ -22,20 +22,39 @@ def folder_convert(folder_list,out_path=None):
 		
 		folder=abspath(folder)
 		print(folder)
-		files = [f for f in listdir(folder) if isfile(join(folder, f))] #hmm I like the more object oriented approach where each of the folder contents has an isFolder attribute; might want a new library
-		out_folder=out_path if out_path else folder # this was previously within file loop, hopefully I'm not missing something
+		#out_folder=out_path if out_path else folder # this was previously within file loop, hopefully I'm not missing something
+		if out_path:
+			out_folder=out_path
+			mkdir(out_folder)
+		else: 
+			out_folder=folder
+			
+		print(out_folder)
+		#files = [f for f in listdir(folder) if isfile(join(folder, f))] #hmm I like the more object oriented approach where each of the folder contents has an isFolder attribute; might want a new library
 		
-		print(files)
-		# loop files
-		for base_fn,ext in files.rsplit('.',1) if '.' in files else (files,''):
-			if ext in extensions_to_convert:
+		for file in listdir(folder):
+			if isfile(join(folder,file)) and '.' in file:
+				print(file)
+				base_fn,ext=file.rsplit('.',1)
+				if ext in extensions_to_convert:
+					command='ffmpeg -i "%s" "%s"' % (join(folder,base_fn+'.'+ext),join(out_folder,base_fn+'.wav'))
+					system(command) #relevant files converted to wav
 
-				command='ffmpeg -i %s %s' % join(folder,base_fn+'.'+ext),join(out_folder,base_fn+'.wav')
-				system(command) #relevant files converted to wav
+				elif ext in extensions_to_keep:
+					# copy the files to the output folder
+					copy2(join(folder,base_fn+'.'+ext),join(out_folder,base_fn+'.'+ext))
 
-			elif ext in extensions_to_keep:
-				# copy the files to the output folder
-				copy2(join(folder,base_fn+'.'+ext),join(out_folder,base_fn+'.'+ext))
+		# print(files)
+		# # loop files
+		# for base_fn,ext in files.rsplit('.',1) if '.' in files else (files,''):
+		# 	if ext in extensions_to_convert:
+
+		# 		command='ffmpeg -i %s %s' % join(folder,base_fn+'.'+ext),join(out_folder,base_fn+'.wav')
+		# 		system(command) #relevant files converted to wav
+
+		# 	elif ext in extensions_to_keep:
+		# 		# copy the files to the output folder
+		# 		copy2(join(folder,base_fn+'.'+ext),join(out_folder,base_fn+'.'+ext))
 
 			#else: do nothing to other types of files
 
@@ -44,6 +63,7 @@ if __name__ == "__main__":
 	import argparse
 	parser=argparse.ArgumentParser()
 	parser.add_argument('folders',nargs='+',help='One or more folders containing audio files to convert')
-	parser.add_argument('-o',nargs=1,dest='out_path',metavar='out_path',default=None,help='Optional out path for "converted" folder')
+	parser.add_argument('-o',dest='out_path',metavar='out_path',default=None,help='Optional out path for "converted" folder')
 	args=parser.parse_args()
+	print(args.out_path)
 	folder_convert(args.folders,args.out_path)
